src/requirements.txt
async-timeout==4.0.3
blinker==1.8.2
click==8.1.7
Flask==3.0.3
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
redis==5.0.8
Werkzeug==3.0.4
zipp==3.20.2
async-timeout==4.0.3
blinker==1.8.2
click==8.1.7
Flask==3.0.3
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
pytz==2024.2
redis==5.0.8
Werkzeug==3.0.4
zipp==3.20.2
